+++ src/__experimental__/components/simple-color-picker/simple-color/simple-color.d.ts
@@ -10,7 +10,7 @@ export interface SimpleColorProps {
   /** Prop for `onBlur` events */
   onBlur?: (ev: React.FocusEvent<HTMLInputElement>) => void;
   /** Prop for `onMouseDown` events */
-  onMouseDown?: (ev: React.SyntheticEvent) => void;
+  onMouseDown?: (ev: React.MouseEvent<HTMLInputElement>) => void;
   /** determines if this color option is selected or unselected */
   checked?: boolean;
   /** [Legacy] Custom classname */