+++ src/components/flat-table/flat-table-cell/flat-table-cell.d.ts
@@ -1,5 +1,6 @@
 import * as React from "react";
 import { PaddingProps } from "styled-system";
+import { TableBorderSize } from "..";
 import * as OptionsHelper from "../../../utils/helpers/options-helper";
 
 export interface FlatTableCellProps extends PaddingProps {
@@ -16,6 +17,10 @@ export interface FlatTableCellProps extends PaddingProps {
   truncate?: boolean;
   /** Title text to display if cell content truncates */
   title?: string;
+  /** Sets a custom vertical right border */
+  verticalBorder?: TableBorderSize;
+  /** Sets the color of the right border */
+  verticalBorderColor?: string;
 }
 
 declare function FlatTableCell(props: FlatTableCellProps): JSX.Element;