+++ src/components/flat-table/flat-table-cell/flat-table-cell.spec.js
@@ -79,4 +79,49 @@ describe("FlatTableRowCell", () => {
       });
     });
   });
+
+  describe.each([
+    ["small", "1px solid #CCD6DB"],
+    ["medium", "2px solid #CCD6DB"],
+    ["large", "4px solid #CCD6DB"],
+  ])(
+    "when the verticalBorder prop is set to %s",
+    (verticalBorder, expectedValue) => {
+      let wrapper;
+
+      it("it overrides the cell border-right size", () => {
+        wrapper = mount(<FlatTableRowCell verticalBorder={verticalBorder} />);
+        assertStyleMatch(
+          {
+            borderRight: expectedValue,
+          },
+          wrapper,
+          { modifier: "&&&" }
+        );
+      });
+    }
+  );
+
+  describe.each([
+    ["goldTint10", "#FFBC1A"],
+    ["#000", "#000"],
+  ])(
+    "when the verticalBorderColor prop is set to %s",
+    (verticalBorderColor, expectedValue) => {
+      let wrapper;
+
+      it("it overrides the cell border-right-color", () => {
+        wrapper = mount(
+          <FlatTableRowCell verticalBorderColor={verticalBorderColor} />
+        );
+        assertStyleMatch(
+          {
+            borderRightColor: expectedValue,
+          },
+          wrapper,
+          { modifier: "&&&" }
+        );
+      });
+    }
+  );
 });