+++ src/components/flat-table/flat-table-header/flat-table-header.component.js
@@ -79,6 +79,10 @@ FlatTableHeader.propTypes = {
    * Callback to report the offsetWidth
    */
   reportCellWidth: PropTypes.func,
+  /** If true sets alternative background color */
+  alternativeBgColor: PropTypes.bool,
+  /** Sets a custom vertical right border */
+  verticalBorder: PropTypes.oneOf(["small", "medium", "large"]),
 };
 
 FlatTableHeader.defaultProps = {